projects/fal_serverless/pyproject.toml
[tool.poetry]
name = "fal_serverless"
version = "0.6.24a0"
description = "fal Serverless is an easy-to-use Serverless Python Framework"
authors = ["Features & Labels <hello@fal.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
auth0-python = "^3.24.0"
requests = "^2.28.1"
isolate = { version = "0.11.0", extras = ["build"] }
grpcio = "^1.50.0"
dill = "0.3.5.1"
isolate-proto = {path = "../isolate_proto", develop = true}
typing-extensions = "4.4"
click = "^8.1.3"
structlog = "^22.3.0"
datadog-api-client = "^2.9.0"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
grpc-interceptor = "^0.15.0"
colorama = "^0.4.6"
portalocker = "^2.7.0"
rich = "^13.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fal-serverless = "fal_serverless.cli:cli"
